fix: keep editorial collapse in existing client module
